Use Narf's exact syscall wire-number type
[hermit2, narf-hermit-design, codex-family-by-ancestry, devbig030, role=impl]

Represent requests with the same unsigned 32-bit wire number Narf snapshots
at trap entry. This removes a lossy signed adapter at the kernel boundary.

Test Plan:
- cargo fmt --all -- --check
- cargo test -p reverie-narf-core -p reverie-narf
- cargo clippy -p reverie-narf-core -p reverie-narf --all-targets -- -D warnings
- cargo +nightly-2025-09-14 check -p reverie-narf-core --target x86_64-unknown-none -Z build-std=core

+++ reverie-narf-core/src/lib.rs
@@ -38,8 +38,8 @@ pub type RawSyscallArgs = [u64; 6];
 /// One explicit native syscall request.
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub struct NarfSyscallRequest {
-    /// Linux syscall number for the current architecture.
-    pub number: i64,
+    /// Exact unsigned syscall wire number for the current architecture.
+    pub number: u32,
     /// Six raw register arguments.
     pub args: RawSyscallArgs,
 }
